using previous refactor on qt as well
