+++ traitsui/testing/tester/qt4/implementation/enum_editor.py
@@ -15,85 +15,40 @@ from traitsui.qt4.enum_editor import (
     SimpleEditor,
 )
 from traitsui.testing.tester import command, locator
+from traitsui.testing.tester.base_classes import _IndexedEditor
 from traitsui.testing.tester.qt4 import helpers
 
 
-class _IndexedListEditor:
-    """ Wrapper for (list) EnumEditor and index """
-
-    def __init__(self, target, index):
-        self.target = target
-        self.index = index
-
-    @classmethod
-    def register(cls, registry):
-        registry.register_solver(
-            target_class=ListEditor,
-            locator_class=locator.Index,
-            solver=lambda wrapper, location: 
-                cls(target=wrapper.target, index=location.index),
-        )
-        registry.register_handler(
-            target_class=cls,
-            interaction_class=command.MouseClick,
-            handler=lambda wrapper, _: helpers.mouse_click_item_view(
+class _IndexedListEditor(_IndexedEditor):
+    target_class = ListEditor
+    handlers = [
+        (command.MouseClick, (lambda wrapper, _: helpers.mouse_click_item_view(
                 model=wrapper.target.target.control.model(),
                 view=wrapper.target.target.control,
-                index=wrapper.target.target.control.model().index(wrapper.target.index, 0),
-            )
+                index=wrapper.target.target.control.model().index(wrapper.target.index, 0)))
         )
+    ]
 
 
-class _IndexedRadioEditor:
-    """ Wrapper for RadioEditor and an index.
-    """
-
-    def __init__(self, target, index):
-        self.target = target
-        self.index = index
-
-    @classmethod
-    def register(cls, registry):
-        registry.register_solver(
-            target_class=RadioEditor,
-            locator_class=locator.Index,
-            solver=lambda wrapper, location:
-                cls(target=wrapper.target,index=location.index),
-        )
-        registry.register_handler(
-            target_class=cls,
-            interaction_class=command.MouseClick,
-            handler=lambda wrapper, _: helpers.mouse_click_qlayout(
+class _IndexedRadioEditor(_IndexedEditor):
+    target_class = RadioEditor
+    handlers = [
+        (command.MouseClick, (lambda wrapper, _: helpers.mouse_click_qlayout(
                 layout=wrapper.target.target.control.layout(),
-                index=wrapper.target.index,
-            )
+                index=wrapper.target.index))
         )
+    ]
 
 
-class _IndexedSimpleEditor:
-    """ Wrapper for (simple) EnumEditor and an index."""
-
-    def __init__(self, target, index):
-        self.target = target
-        self.index = index
-
-    @classmethod
-    def register(cls, registry):
-        registry.register_solver(
-            target_class=SimpleEditor,
-            locator_class=locator.Index,
-            solver=lambda wrapper, location:
-                cls(target=wrapper.target, index=location.index),
+class _IndexedSimpleEditor(_IndexedEditor):
+    target_class = SimpleEditor
+    handlers = [
+        (command.MouseClick, (lambda wrapper, _: helpers.mouse_click_combobox(
+            combobox=wrapper.target.target.control,
+            index=wrapper.target.index,
+            delay=wrapper.delay))
         )
-        registry.register_handler(
-            target_class=cls,
-            interaction_class=command.MouseClick,
-            handler=lambda wrapper, _: helpers.mouse_click_combobox(
-                combobox=wrapper.target.target.control,
-                index=wrapper.target.index,
-                delay=wrapper.delay,
-            ),
-        )
+    ]
 
 
 def register(registry):
